WIP: Delete Associations in the background Motivation:
In #36912 I added background destroys, this pr brings in deletes as
well. Currently in the GitHub codebase we have methods that make both of
these functions available, I just learned this thanks @seejohnrun.
However it is hidden and likely underutlized due to the fact that it is
not core Rails functionality.

It would be nice to upstream both these tools if the Rails core team is
interested.

Related Issues:

- #34901
- rails/actionmailbox#16
- #36912

Changes:

- Added a new job to handle deleting associations.
- Added a new dependent option on associations,
background_delete (Open to other names).

+++ activerecord/lib/active_record/associations/belongs_to_association.rb
@@ -11,6 +11,8 @@ module ActiveRecord
         when :destroy
           target.destroy
           raise ActiveRecord::Rollback unless target.destroyed?
+        when :background_delete
+          ActiveRecord::DeleteAssociationLaterJob.perform_later([target])
         else
           target.send(options[:dependent])
         end
